Cleanup in models
The technology and the port are not needed as structures.
The regex in the technology will be replaced by the notion of
checkers, structures created to check for one technology. They
will contain the regex and do the job. The technology name
has been moved in the finding, and the version in the finding
is now optional.