Create a running application
A module "application" has been created. It contains a structure to
manage the global execution of the program.
The IP or hostname and the port are read from argv.
The printing of findings is now done by a writer. By doing so, it's
possible to export the finding in a different way in the future.

Usage: ./sanca <ip_hostname> <port>