Improve banner conversion to string
Banners fetched via TCP are converted to string to be parsed by the
checkers. The method used was returning a Result<&str>, because if an
invalid character was found, the banner was considered as invalid utf-8.
The new method is "lossy", that means all utf-8 characters are read, and
invalids one are replaced by the "invalid" character. By doing so,
invalid characters can just be ignored, and valid ones can be used
by checkers.