Use clap and implement a better structure
Clap is used to have command line arguments managed easily.
Sanca can receive the scan type, the ip or hostname, the port, and
the URL. Use sanca --help to learn more.

The method Application::run() is only an orchestrator, and the actual
scan is done in Application::tcp_udp_scan(). HTTP scans will be done
is a separate method.
The ip or hostname is not printed during the scan anymore, the data
has been provided to the writer, and it prints all the data.