Add the concept of Technology
In addition to providing a scan type, it's now possible to provide
technologies to check for.
Each checker, in addition to be bound to scan types, are also bound to
one technology. For example, the OpenSSH checker supports the TCP scan,
and the OpenSSH technology.

By doing so it's possible to filter the checker that are used and
avoid to call checkers that don't match the scan type.
In the future it will be possible to configure the behaviour of the
readers based on the technology. For example to check for PHP it could
be enough to check at the provided URL and then at /phpinfo.php, without
fetching the JavaScript files found on the original page. But to
detect JavaScript libraries, the behaviour has to be different.