Reimplement UrlRequest::from_technologies()
The method UrlRequest::from_technologies() takes a list of
technologies and returns a list of UrlRequest.
The method was partially implemented, enough for working, but
it was possible to have duplicate UrlRequests, two at the same
URL but with a different value of fetch_js. It would have been
annoying to send more requests than needed.

The new implementation uses HashMap instead of Vec, so when inserting
the URL it's easier to manipulate the item without inserting a duplicate.