Implement HttpReader
HttpReader is now able to send HTTP requests and return Vec<UrlResponse>.
It has the following logic:

1. It takes a list of UrlRequest to handle. Each UrlRequest is a page to load.
2. For each UrlRequest it fetches the URL.
3. Fetching of JavaScript files is not implmented yet, but all
   files will be downloaded in parallel.
4. All UrlRequests are handled in parallel using Futures and async/.await.

The Application calls the HttpReader and waits (read block_on()) until all
UrlResponses are ready.
It should be the only part of Sanca needing asynchronous programming,
the checkers and writers will work in procedural way.
the whole pool of UrlRequests has been handled.