Checkers can have several regexes
Checkers used to have only one regex, but it is not enough because
a given technology (e.g. the operating system) could be identified
in different scan types, in a different way.
When scanning the OpenSSH banner, the operating system can be identified.
But it can also be identified by looking at HTTP headers.

To be able to handle these different cases, each checker has a HashMap
of regexes. Each regex has a name and can be accessed in a clear way.