Improve headers handling in checkers
HttpCheckers often read some HTTP headers, so they need to filter
the headers from the UrlResponse, check if the headers they're looking
for are present, and if so, they can handle them.

Instead of doing this manually in each HttpChecker, a method has been
added in urlResponse to filter the headers and return the needed headers.