Check always the main URL first
The initial order of UrlRequest with the main URL being the first was
not preserved, and the order of UrlResponse is the same. So, checkers
used to check any URL at first.

It is preferred to check first on the main page if a technology matches,
and if not present it's OK to check on other ones. By doing so it's easier
to understand, it provided a URL and at this URL we found a technology.