Stop checking technologies against ther servers
Some technologies like PHP, should be identified using only the original
server. If a JavaScript file is hosted on a CDN, we shouldn't check the
HTTP headers when fetching the file. We could identify the PHP version
of the remote server, and report the wrong version for the URL we were
supposed to check.