Improve OS detection
OSChecker first checks in HTTP headers, and then in the body.
If nothing was found in the headers and the body is a
Nginx not found page, the OS was not detected.

Now the OS & optionally its version are identified even when
looking at the body of a Nginx page.