Stop using the query string from main URL in URL generation
The query string, if provided in the main URL, was applied to all
the UrlRequests, even the JavaScript files.

It's not needed, and in some cases it could cause problems.
If a JavaScript file is found at "/this.js?a=1&b=2", the query string
would be used but the main URL's query string would be applied also,
like "?c=3&d=4".

Now, the query string of the main URL is only used for the original
request, and the other requests use their own, if they have one.