Add assertion failure messages in check_finding()
This function is a utility for all unit tests in the "checkers"
module. It avoids to recreate this function in each checker's
"tests" submodule. The counterpart is when an assertion failed,
the message are less explicit since the error doesn't occur
in the exact file of the test. It's possible to get more information
by setting the RUST_BACKTRACE environment variable to 1. To help
even more, messages have been added when an assertion fails, it
should be easier to debug since they prints the expected values
and the ones we got.