Add submodules in the models module
The "models" module was only one file and all the models were defined
inside. The file was big, contained a lot of things, and it would
be even bigger once the unit tests added in it.

Now the models module looks like this:

models/
  mod.rs
  technology.rs
  reqres.rs

The technology module contains the Technology enum and the implemented methods,
and the reqres module contains everything related to HTTP requests & responses.
So, the logic is separated and it will be cleaner.