Send the right MIME type when fetching JavaScript
In the HTTP request header "Accept", the MIME type was always "text/html".
It's correct only for default requests, but when fetching JavaScript
files it may be required by the server to send the right MIME type,
application/javascript.
Now the right MIME type is sent according to the request type.