Fixed OSChecker when OpenSSH wasn't installed as a package
When OpenSSH isn't installed as a package, its banner doesn't contain
the name of the OS. In this case, OSChecker used to panick.
The OS part in the banner was optional in OSChecker, it was a mistake.
Now it's mandatory to be detected. It doesn't affect the detection of
OpenSSH itself.