Improve OS version detection from OpenSSH
At the end of the OpenSSH banner it's possible to have a debX or bpoY
where X and Y are the OS version. When available, they are used to
read the OS version. It works for "normal" and backports repositories.
The OS version deduction is used only when the deb / bpo is not
available.