Give the whole application::Args structure to the writers
The writers used to get the ip_hostname, the port and the url as
parameters. It was a constraint since it wasn't possible to provide
more parameters to a given writer, like a filename to write the
output.

Now, the writers take the whole application::Args structure as
parameter, so they can use whatever they need. It's also cleaner.