Fix the reported URL of finding if a redirection occured
The HttpReader used to use the URL coming from the request, when
filling the URL of finding. In most cases it's fine, but when the
initial request gets a redirection, it leads to a false URL being
reported.

The HttpReader now reads the URL from the response, so it can be sure
that the reported URL is the right one.