Improve MIME type handling
Some servers strictly respect the Accept header, and refuse to return the
resource at the given URL if its MIME type isn't listed in Accept.

The UrlRequestType was used to determine which MIME type to use, but it
was not perfect. It was using Accept: text/html for the main URL, which
is wrong when the main URL points to a JavaScript file.

Now, the Accept header looks like this in all cases:
Accept: text/html,application/javascript,*/*;q=0.8