Technology version can be splitted in several parts
In most cases, the version is written as in "1.2.3", but it's sometimes,
it's splitted on several places of a file. For example, TinyMCE uses
something similar to "majorVersion:'1',minorVersion:'2.3'".

Now, the version can be splitted in the regular expressions. When the
version is not splitted, it's enough to take it all in the matching
group "version1", but when it's splitted it's possible to use the
groups "version2", "version3" up to "version5".
The version parts are joined on a dot.