TcpChecker handles findings in the same way than HttpChecker
HttpChecker used extract_finding_from_captures() to automatically
extract a finding from the matching groups found in the regular
expression.

This behaviour has been generalized and is now used by the TcpChecker
too.
This way, it's easier to extract a finding when adding a new technology,
and the internal API is more consistent.

Only OSChecker doesn't use extract_finding_from_captures() because it's
unique, it searches in other technologies' banners to identify its own
version.