Improve regexes handling in checkers
Each checker has one or several regexes to match input against. Only a
part of the whole match is kept as an evidence, because it could be
very long.
The number of chars to preserve at the beginning and the end of the
whole match were normal values passed to a method call. Since these
values depend on each regex, they have been attached to them in
tuples.
Instead of storing a HashMap<&str, Regex>, the checkers
now store a HashMap<&str, (Regex, usize, usize)>. The first usize
is the number of chars to keep at the beginning of the whole match
for this regex, and the second usize is the number of chars to keep
at the end of the whole match for this regex.

It makes the code cleaner because these values are now parts of the
checkers. Another benefit is that the checkers searching in only one
file, like the ones related to JavaScript libraries, now check for all
their regexes in a loop instead of one by one, and a lot of code has
been removed.