Capture the whole OpenSSH banner in regex
The TcpCheckers no longer use the whole banner in the evidence if the
regex matches, no they use the wholematch capture group in the regex.

The OpenSSH banner regex was missing some allowed chars at the end,
so the end of the banner wasn't captured. Example, with this banner:

SSH-2.0-OpenSSH_9.2p1 Debian-2+deb12u1

The following part wasn't captured in the evidence.

-2+deb12u1