Don't truncate the evidence for MySQL & MariaDB
Their banners contain non-printable characters which are replaced by
the replacement character (U+FFFD). When the evidence is truncated
on one of these chars, the program panic because the character is
multi-bytes wide.
To prevent this, the evidence isn't truncated. To be accurate,
the evidence can be up to 200 chars without being truncated. It
should be enough.