Extracted UrlRequest have fetch_js to true in HttpReader
It was set to false because we don't need to extract JavaScript URLs
from a JavaScript file, it's needed only on HTML files.
Setting it to true in HttpReader::extract_urls() gives the right request
type to the UrlResponse, which is used by the checkers. Some server-side
technologies checkers only check on UrlResponse with request type Default.