Improve vulnerabilities storing
The CVEs of technologies can be stored in cache using JSON files. The
JSON wasn't pretty, difficult to read and to edit. It's now formatted
on several lines.
Moreover, the base score of the CVE was not accurate, 6.1 became 6.098374
so the type had to be changed to f64 instead of f32.